create a base class so limit test can run separately
