+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/JsonRpcConfigurationTest.java
@@ -33,7 +33,7 @@ public class JsonRpcConfigurationTest {
     assertThat(configuration.getCorsAllowedDomains()).isEmpty();
     assertThat(configuration.getRpcApis())
         .containsExactlyInAnyOrder(RpcApis.ETH, RpcApis.NET, RpcApis.WEB3);
-    assertThat(configuration.getMaxActiveConnections()).isEqualTo(20);
+    assertThat(configuration.getMaxActiveConnections()).isEqualTo(80);
   }
 
   @Test
